Add coverage directory for coverage analysis
If you install xdebug locally, you can run phpunit with coverage
analysis to verify what parts of your code the tests is running. Here is
my command line for runnning tests for the member class

```
./phpunit tests/MemberTest.php --colors=always --verbose \
--coverage-html tests/coverage --whitelist classes/member.class.php
```

The first argument is what test files to run. If it is a directory all
tests in that directory will run. If it is a file only that file will be
run.

The `--coverage-html <path>` tells phpunit to do a coverage analysis
using xdebug and store the html files in `<path>`. But for this to work
we have to whitelist what files to include in the coverage analysis
which is what the last flag does.