Fix signup form token validation.
This commit brings in javascript functions to replace existing tokens.
New tokens are only returned if the user passed the first token
validation. So if the user did not pass validation they WILL NOT receive
a new, valid token.

As some of us discussed: It should not make a difference if the
generated token is sent to the user as HTML in a new page response or as a
JSON response to an ajax request.